remove sctp dep
