+++ package.json
@@ -123,7 +123,6 @@
 		"open-cli": "^8.0.0",
 		"pick-port": "^2.2.0",
 		"prettier": "^3.8.1",
-		"sctp": "^1.0.0",
 		"ts-jest": "^29.4.6",
 		"typescript": "^5.9.3",
 		"typescript-eslint": "^8.55.0",
